k8s/telemetry/oob-collectors/bmc-sol-collector/code/pip-requirements.txt
certifi==2020.12.5
chardet==4.0.0
gpg==1.10.0
iniparse==0.5
pexpect==4.8.0
prometheus-client==0.9.0
ptyprocess==0.7.0
pyaml==20.4.0
PyYAML==5.3.1
requests==2.25.1
urllib3==1.26.2
websocket-client==0.57.0
